Separate the SiteData into two parts to avoid polluting the setuptool with MPI
