+++ Code/geometry/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_library(
 	hemelb_geometry BlockTraverser.cc BlockTraverserWithVisitedBlockTracker.cc 
-	GeometryReader.cc needs/Needs.cc LatticeData.cc SiteData.cc
+	GeometryReader.cc needs/Needs.cc LatticeData.cc SiteDataBare.cc SiteData.cc
 	SiteTraverser.cc VolumeTraverser.cc Block.cc 
 	decomposition/BasicDecomposition.cc decomposition/OptimisedDecomposition.cc
 	neighbouring/NeighbouringLatticeData.cc	neighbouring/NeighbouringDataManager.cc
